feat: some updates to earn page
