+++ components/Currency/CurrencyAmount/CurrencyAmount.tsx
@@ -5,7 +5,7 @@ import { formatCurrency, NumericValue, toBigNumber } from 'utils/formatters/numb
 
 import { CurrencyKey } from 'constants/currency';
 
-import { ContainerRowMixin } from '../common';
+import { ContainerRow } from '../common';
 
 type CurrencyAmountProps = {
 	amountCurrencyKey: CurrencyKey;
@@ -37,9 +37,7 @@ export const CurrencyAmount: FC<CurrencyAmountProps> = ({
 	</Container>
 );
 
-const Container = styled.span`
-	${ContainerRowMixin};
-`;
+const Container = styled(ContainerRow)``;
 
 const Amount = styled.span`
 	color: ${(props) => props.theme.colors.white};
