+++ components/Currency/common.tsx
@@ -1,4 +1,12 @@
-export const ContainerRowMixin = `
-    display: inline-grid;
-    grid-gap: 1px;
+import styled from 'styled-components';
+import media from 'styles/media';
+
+export const ContainerRow = styled.span`
+	display: inline-grid;
+	grid-gap: 1rem;
+
+	${media.lessThan('md')`
+        display: flex;
+        flex-direction: row;
+    `}
 `;
